// web/src/api/model/jobStatus.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Brisa API
 * API for the Brisa Invoice Scraper
 * OpenAPI spec version: 1.0.0
 */

export type JobStatus = (typeof JobStatus)[keyof typeof JobStatus];

export const JobStatus = {
  created: 'created',
  running: 'running',
  waiting_captcha: 'waiting_captcha',
  completed: 'completed',
  failed: 'failed',
} as const;